Refactored specie model to integrate spatial attributes and settings.  Refactored edit specie and edit parameter to accomadate new specie model design.  Removed edit-reaction-var.  Set up spatial-settings, nospatial-settings, and initial-conditions model for spatial implementation.
