+++ src/models/non-spatial-specie.js
@@ -1,0 +1,11 @@
+var State = require('ampersand-state');
+
+module.exports = State.extend({
+  props: {
+    value: 'number',
+    mode: {
+      type: 'string',
+      default: 'dynamic'
+    }
+  }
+});
