+++ src/models/spatial-specie.js
@@ -1,0 +1,11 @@
+var State = require('ampersand-state');
+var InitialCondition = require('./initial-condition')
+
+module.exports = State.extend({
+  props: {
+    diffusionCoeff: 'number'
+  },
+  children: {
+    initialCondition: InitialCondition
+  }
+})
